refactor: common 'setup_router' function in artifact routes module
